move todoButtons to it's own component
