+++ src/components/todoItem/todoItem.scss
@@ -1,45 +1,6 @@
 @import "styles/colors";
 @import "styles/variables";
 
-.item_card {
-  background-color: $primary-white;
-  border: $sp-1 solid $primary-border;
-  height: calc($sp-24 * 5.1);
-  padding-left: $sp-12;
-  position: relative;
-  width: calc($sp-36 * 8);
-
-  &__bottom {
-    bottom: $sp-4;
-    position: absolute;
-    width: calc($sp-36 * 8);
-  }
-
-  &__actions {
-    display: flex;
-    gap: $sp-12;
-    align-items: center;
-
-    &__img {
-      height: $sp-20;
-      cursor: pointer;
-    }
-
-    &--completed {
-      display: flex;
-      align-items: center;
-      justify-content: space-between;
-      width: calc($sp-36 * 7);
-    }
-  }
-
-  &__time {
-    color: $primary-gray;
-    font-size: $sp-12;
-    font-weight: 700;
-  }
-}
-
 .completed {
   color: $primary-green;
   text-decoration: line-through;
